net, localnet: Move IPAM tests to dedicated sub-package
Create tests/network/localnet/ipam/ sub-package to organize
IPAM-specific tests and fixtures separately from the main
localnet tests.
